Wed, Aug 2, 2023, 7:01 PM +04:00

+++ hours/hours.c
@@ -42,6 +42,6 @@ float calc_hours(int hours[], int weeks, char output)
     else
     {
         int average = (float)total / weeks;
-        return average;
+        return (float)average;
     }
 }
